wip: add broadcast handler for renotify queue
