+++ packages/store/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./database";
+export * from "./change_classifier"
